+++ lib/screens/home_page/editor_pane/details_card/request_pane/request_body.dart
@@ -5,24 +5,18 @@ import 'package:apidash/widgets/widgets.dart';
 import 'package:apidash/consts.dart';
 import 'request_form_data.dart';
 
-class EditRequestBody extends ConsumerStatefulWidget {
+class EditRequestBody extends ConsumerWidget {
   const EditRequestBody({super.key});
 
   @override
-  ConsumerState<EditRequestBody> createState() => _EditRequestBodyState();
-}
-
-class _EditRequestBodyState extends ConsumerState<EditRequestBody> {
-  @override
-  Widget build(BuildContext context) {
-    final activeId = ref.watch(activeIdStateProvider);
+  Widget build(BuildContext context, WidgetRef ref) {
+    final selectedId = ref.watch(selectedIdStateProvider);
     final requestModel = ref
         .read(collectionStateNotifierProvider.notifier)
-        .getRequestModel(activeId!);
-    ContentType? requestBodyStateWatcher = (ref
-            .watch(collectionStateNotifierProvider)![activeId]
-            ?.requestBodyContentType) ??
-        ContentType.values.first;
+        .getRequestModel(selectedId!);
+    final contentType = ref.watch(selectedRequestModelProvider
+        .select((value) => value?.requestBodyContentType));
+
     return Container(
       decoration: BoxDecoration(
         color: Theme.of(context).colorScheme.background,
@@ -43,18 +37,30 @@ class _EditRequestBodyState extends ConsumerState<EditRequestBody> {
             ),
           ),
           Expanded(
-            child: requestBodyStateWatcher == ContentType.formdata
-                ? const FormDataWidget()
-                : TextFieldEditor(
-                    key: Key("$activeId-body"),
-                    fieldKey: "$activeId-body-editor",
-                    initialValue: requestModel?.requestBody,
-                    onChanged: (String value) {
-                      ref
-                          .read(collectionStateNotifierProvider.notifier)
-                          .update(activeId, requestBody: value);
-                    },
-                  ),
+            child: switch (contentType) {
+              ContentType.formdata => const FormDataWidget(),
+              // TODO: Fix JsonTextFieldEditor & plug it here
+              ContentType.json => TextFieldEditor(
+                  key: Key("$selectedId-json-body"),
+                  fieldKey: "$selectedId-json-body-editor",
+                  initialValue: requestModel?.requestBody,
+                  onChanged: (String value) {
+                    ref
+                        .read(collectionStateNotifierProvider.notifier)
+                        .update(selectedId, requestBody: value);
+                  },
+                ),
+              _ => TextFieldEditor(
+                  key: Key("$selectedId-body"),
+                  fieldKey: "$selectedId-body-editor",
+                  initialValue: requestModel?.requestBody,
+                  onChanged: (String value) {
+                    ref
+                        .read(collectionStateNotifierProvider.notifier)
+                        .update(selectedId, requestBody: value);
+                  },
+                ),
+            },
           )
         ],
       ),
@@ -62,29 +68,22 @@ class _EditRequestBodyState extends ConsumerState<EditRequestBody> {
   }
 }
 
-class DropdownButtonBodyContentType extends ConsumerStatefulWidget {
+class DropdownButtonBodyContentType extends ConsumerWidget {
   const DropdownButtonBodyContentType({
     super.key,
   });
 
   @override
-  ConsumerState<DropdownButtonBodyContentType> createState() =>
-      _DropdownButtonBodyContentTypeState();
-}
-
-class _DropdownButtonBodyContentTypeState
-    extends ConsumerState<DropdownButtonBodyContentType> {
-  @override
-  Widget build(BuildContext context) {
-    final activeId = ref.watch(activeIdStateProvider);
-    final requestBodyContentType = ref.watch(activeRequestModelProvider
+  Widget build(BuildContext context, WidgetRef ref) {
+    final selectedId = ref.watch(selectedIdStateProvider);
+    final requestBodyContentType = ref.watch(selectedRequestModelProvider
         .select((value) => value?.requestBodyContentType));
     return DropdownButtonContentType(
       contentType: requestBodyContentType,
       onChanged: (ContentType? value) {
         ref
             .read(collectionStateNotifierProvider.notifier)
-            .update(activeId!, requestBodyContentType: value);
+            .update(selectedId!, requestBodyContentType: value);
       },
     );
   }