+++ test/widgets/response_widgets_test.dart
@@ -68,7 +68,7 @@ void main() {
         theme: kThemeDataLight,
         home: const Scaffold(
           body: ResponseTabView(
-              activeId: '1', children: [Text('first'), Text('second')]),
+              selectedId: '1', children: [Text('first'), Text('second')]),
         ),
       ),
     );
@@ -229,7 +229,7 @@ void main() {
         title: 'Response Body',
         theme: kThemeDataLight,
         home: Scaffold(
-          body: ResponseBody(activeRequestModel: requestModelNoResponseBody),
+          body: ResponseBody(selectedRequestModel: requestModelNoResponseBody),
         ),
       ),
     );
@@ -254,7 +254,8 @@ void main() {
         title: 'Response Body',
         theme: kThemeDataLight,
         home: Scaffold(
-          body: ResponseBody(activeRequestModel: requestModelNoResponseHeaders),
+          body:
+              ResponseBody(selectedRequestModel: requestModelNoResponseHeaders),
         ),
       ),
     );
@@ -281,7 +282,8 @@ void main() {
         title: 'Response Body',
         theme: kThemeDataLight,
         home: Scaffold(
-          body: ResponseBody(activeRequestModel: requestModelNoResponseHeaders),
+          body:
+              ResponseBody(selectedRequestModel: requestModelNoResponseHeaders),
         ),
       ),
     );
@@ -308,7 +310,8 @@ void main() {
         title: 'Response Body',
         theme: kThemeDataLight,
         home: Scaffold(
-          body: ResponseBody(activeRequestModel: requestModelNoResponseHeaders),
+          body:
+              ResponseBody(selectedRequestModel: requestModelNoResponseHeaders),
         ),
       ),
     );
@@ -431,7 +434,9 @@ void main() async {
     expect(find.byType(Image), findsOneWidget);
   });
 
-  testWidgets('Testing Body Success tap segment', (tester) async {
+  testWidgets(
+      'Testing Body Success tap segment. formattedBody is always shown in Raw',
+      (tester) async {
     await tester.pumpWidget(
       MaterialApp(
         title: 'Body Success',
@@ -458,7 +463,37 @@ void main() async {
 
     await tester.tap(find.text('Raw'));
     await tester.pumpAndSettle();
-    expect(find.text('Formatted Hello from API Dash'), findsNothing);
+    expect(find.text('Formatted Hello from API Dash'), findsOneWidget);
+    expect(find.text('Raw Hello from API Dash'), findsNothing);
+  });
+
+  testWidgets('Testing Body Success tap segment for formattedBody null',
+      (tester) async {
+    await tester.pumpWidget(
+      MaterialApp(
+        title: 'Body Success',
+        theme: kThemeDataLight,
+        home: Scaffold(
+          body: BodySuccess(
+            body: 'Raw Hello from API Dash',
+            formattedBody: null,
+            mediaType: MediaType("text", "csv"),
+            options: const [
+              ResponseBodyView.code,
+              ResponseBodyView.raw,
+            ],
+            bytes: kBodyBytesJpeg,
+            highlightLanguage: 'txt',
+          ),
+        ),
+      ),
+    );
+
+    await tester.pumpAndSettle();
+    expect(find.text('Raw Hello from API Dash'), findsOneWidget);
+
+    await tester.tap(find.text('Raw'));
+    await tester.pumpAndSettle();
     expect(find.text('Raw Hello from API Dash'), findsOneWidget);
   });
 }